chatbot/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langchain==0.3.25
langchain-google-genai==2.1.4
langchain-community==0.3.24
langchain-text-splitters==0.3.8
chromadb==1.0.7
python-dotenv==1.1.0
